requirements.txt


python-telegram-bot[job-queue]==20.7
google-generativeai==0.8.3
python-dotenv
